fix(catalog): make the Lo-fi grit display independent of FMA
At 0.23, 1 + 0.23f * 15 falls exactly between two float32 values, so the
displayed tenth depended on whether the compiler fused the multiply-add:
Apple clang printed 4.5x, the Linux CI build 4.4x. A fixture regenerated
on macOS therefore failed the manager parity test in CI. Compute the value
in double (4.45000006, clear of the tie) on the device and in the
TypeScript mirror; both now give 4.5x with and without contraction.

Also cast float products to double before accumulating them in the new
modulation tests (CodeQL: multiplication result converted to larger type).

+++ apps/manager/src/effects/daisyValues.ts
@@ -251,7 +251,10 @@ function delayDisplay(mode: string, key: string): NumberDisplay {
     if (mode === "filter") return choices(["Low-pass", "Band-pass", "High-pass"]);
     if (mode === "pattern") return choices(["Straight", "Dotted 8th", "Triplet"]);
     if (mode === "lofi") return custom((value) => {
-      const flutter = Math.fround(1 + Math.fround(Math.fround(clamp(value)) * 15));
+      // Mirrors the device: the float32 knob value, then double arithmetic,
+      // then float32 for printing. The all-float32 form sat on a rounding tie
+      // at 0.23 whose result depended on fused multiply-add.
+      const flutter = Math.fround(1 + Math.fround(clamp(value)) * 15);
       return `${16 - Math.trunc(clamp(value) * 12)} bit / ${number(flutter, 1)}x`;
     });
     if (mode === "swell") return custom((value) => `${number(20 * Math.log10(0.05 + clamp(value) * 0.2), 1)} dBFS`);
